added a function see description
to be able to quickly set 15 winners to local storage. no need to keep it but it is handy to have if you wanna show a nice highscore list without having to play the game 10 times... just uncomment from line 34 to 58. but remember to comment it out again once you loaded the game page.